feat: move audio playing function into offscreen page.
