+++ src/background/index.ts
@@ -1,9 +1,5 @@
-import { audioElement, stopPlayback, readText, readPage } from "./ttsService";
+import { setupOffscreenDocument } from "../offScreen/setup";
 // src/background/index.ts
-
-chrome.runtime.onStartup.addListener( () => {
-  console.log(`onStartup()`);
-});
 
 chrome.runtime.onInstalled.addListener(() => {
   // Add context menu for reading selected text
@@ -21,22 +17,77 @@ chrome.runtime.onInstalled.addListener(() => {
   });
 });
 
-chrome.contextMenus.onClicked.addListener((info, tab) => {
+export async function grabAndReadPage( ) {
+  await setupOffscreenDocument('offScreen/off-screen.html')
+  
+  chrome.tabs.query({ active: true, currentWindow: true }, async (tabs) => {
+    if (!tabs || tabs.length === 0 || !tabs[0].id) return;
+    
+    const tabId = tabs[0].id;
+    const settings = await chrome.storage.sync.get({
+      voiceName: "en-US-ChristopherNeural",
+      customVoice: "",
+      speed: 1.2,
+    });
+
+    chrome.scripting.executeScript({
+      target: { tabId: tabId },
+      func: () => document.body.innerText
+    }).then((results) => {
+      if (results && results.length > 0 && results[0].result) {
+        const pageContent = results[0].result;
+        console.log("page content :", pageContent);
+        chrome.runtime.sendMessage({
+          type: 'readText',
+          target: 'offscreen',
+          data: pageContent,
+          settings
+        });
+      }
+    });
+  });
+}
+
+
+chrome.contextMenus.onClicked.addListener(async (info, tab) => {
+
+  if (!tab.id ) {
+    console.error('Tab Not Found.')
+    return 
+  }
+
+  const settings = await chrome.storage.sync.get({
+    voiceName: "en-US-ChristopherNeural",
+    customVoice: "",
+    speed: 1.2,
+  });
+  
+
   if (
     info.menuItemId === "readAloud" &&
-    info.selectionText &&
-    tab?.id !== undefined
+    info.selectionText
   ) {
-    // Handle reading selected text
-    readText(info.selectionText);
+    await setupOffscreenDocument('offScreen/off-screen.html')
+    chrome.runtime.sendMessage({
+      type: 'readText',
+      target: 'offscreen',
+      data: info.selectionText,
+      settings
+    });
   } else if (info.menuItemId === "readPage" && tab?.id !== undefined) {
-    readPage();
+    grabAndReadPage()
   }
 });
 
-chrome.commands.onCommand.addListener((command) => {
+chrome.commands.onCommand.addListener(async (command) => {
   if (command === "read-selected-text") {
-    console.log("Executing read-selected-text command");
+    await setupOffscreenDocument('offScreen/off-screen.html')
+
+    const settings = await chrome.storage.sync.get({
+      voiceName: "en-US-ChristopherNeural",
+      customVoice: "",
+      speed: 1.2,
+    });
 
     chrome.tabs.query({ active: true, currentWindow: true }, (tabs) => {
       if (!tabs || tabs.length === 0 || !tabs[0].id) return;
@@ -52,8 +103,14 @@ chrome.commands.onCommand.addListener((command) => {
           if (results && results.length > 0 && results[0].result) {
             const selectedText = results[0].result;
             console.log("Selected text:", selectedText);
-
-            readText(selectedText);
+            
+            // Send message to offscreen document
+            chrome.runtime.sendMessage({
+              type: 'readText',
+              target: 'offscreen',
+              data: selectedText,
+              settings
+            });
           } else {
             console.log("No text selected.");
           }
@@ -63,29 +120,44 @@ chrome.commands.onCommand.addListener((command) => {
   }
 });
 
-chrome.runtime.onMessage.addListener((message, sender, sendResponse) => {
-  switch (message.action) {
-    case "background:togglePause":
-      console.log("Received pause toggle:", message.isPaused);
-
-      if (message.isPaused) {
-        audioElement?.pause();
-        chrome.runtime.sendMessage({
-          action: "controlPanel:updatePause",
-          isPaused: true,
-        });
-      } else {
-        audioElement?.play();
-        chrome.runtime.sendMessage({
-          action: "controlPanel:updatePause",
-          isPaused: false,
-        });
-      }
-      break;
-
-    case "background:stopPlayback":
-      console.log("Stopping playback and removing control panel.");
-      stopPlayback();
-      break;
+chrome.runtime.onMessage.addListener(async (message, sender, sendResponse) => {
+  
+  // forward messages to offscreen document
+  if (message.action === "offscreen:readPage") {
+    grabAndReadPage()
+  }
+  else if (message.action === "offscreen:togglePause") {
+    console.log("Received pause toggle:", message.isPaused);
+    chrome.runtime.sendMessage({
+      type: 'togglePause',
+      target: 'offscreen',
+      isPaused: message.isPaused
+    });
+  } else if (message.action === "offscreen:stopPlayback") {
+    console.log("Stopping playback and removing control panel.");
+    chrome.runtime.sendMessage({
+      type: 'stopPlayback',
+      target: 'offscreen',
+    });
+  }
+  // forward messages to content script
+  if (message.action ==='controlPanel:updatePause') {
+    chrome.runtime.sendMessage({
+      action: 'controlPanel:updatePause',
+      isPaused: message.isPaused
+    });
+  }else if (message.action ==='controlPanel:remove') {
+    chrome.runtime.sendMessage({
+      action: 'controlPanel:remove',
+    });
+  }else if (message.action ==='controlPanel:create') {
+    chrome.runtime.sendMessage({
+      action: 'controlPanel:create',
+    });
+  }else if (message.action ==='controlPanel:updateLoading') {
+    chrome.runtime.sendMessage({
+      action: 'controlPanel:updateLoading',
+      isLoading: message.isLoading
+    });
   }
 });
