+++ src/offScreen/off-screen.html
@@ -1,0 +1,10 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <title>Offscreen Audio</title>
+  </head>
+  <body>
+    <script src="bundle.js"></script>
+  </body>
+</html>
